Change ints to longs, to avoid integer overflow in large arrays
please see #3

+++ src/mpd/SerialMinimumPairwiseDistance.java
@@ -3,8 +3,8 @@ package mpd;
 public class SerialMinimumPairwiseDistance implements MinimumPairwiseDistance {
 
     @Override
-    public long minimumPairwiseDistance(int[] values) {
-        long result = Integer.MAX_VALUE;
+    public long minimumPairwiseDistance(long[] values) {
+        long result = Long.MAX_VALUE;
         for (int i = 0; i < values.length; ++i) {
             for (int j = 0; j < i; ++j) {
                 // Gives us all the pairs (i, j) where 0 <= j < i < values.length
